pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.0.1"
dependencies = [
  "make87[zenoh,protobuf]==0.1.0",
  "make87-messages-pb5",
  "fastapi",
  "uvicorn[standard]"
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.9,<3.13"


[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["**/*.html"]
